Consolidated all plotting and summarising to one file
